Add in install
Add in the install and install subclasses along with tests. These are
still not completely fleshed out and local testing is getting a weird
dependency cycle on the python::virtualenv call. This commit is partly
to get the code committed, but also to see if Travis-CI has an issue
with it too as I've been having some weird issues with my spec test
environment