Do not force a gerrit default config
The configuration file required sections change between the current
released pip version and what's in dev. As having a [gerrit] section
causes issues if you also have the new [connection gerrit] we're just
going to drop the default gerrit defintion. This does have the side
effect of forcing users to pass in a config_override which they should
be anyway.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>